+++ CHANGELOG.md
@@ -6,11 +6,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/bohrsty/agenda-display/compare/v0.1.0...HEAD)
 ### Added
+- settings ui (timeout and caldav connections)
+- settings api endpoints
+- check for settings json in redis
+- api exception
+- nanoid to generate ids
 - agenda components
 - date-fns for date manipulation
 - api endpoint for agenda (data mocked)
 
 ### Changed
+- redis image to redis-stack image (support for json module)
 - LocalisationContext to provide date-fns locale
 
 ## [v0.1.0](https://github.com/bohrsty/agenda-display/compare/81c0b52...v0.1.0)